+++ eelbrain/_wxutils/ID.py
@@ -95,3 +95,5 @@ UNDO = wx.NewId()
 
 WINDOW_MINIMIZE = wx.NewId()
 WINDOW_ZOOM = wx.NewId()
+
+YIELD_TO_TERMINAL = wx.NewId()